@@ -1,6 +1,6 @@
 {
   "name": "function-bind",
-  "version": "1.1.1",
+  "version": "1.1.2",
   "description": "Implementation of Function.prototype.bind",
   "keywords": [
     "function",
@@ -9,7 +9,13 @@
     "es5"
   ],
   "author": "Raynos <raynos2@gmail.com>",
-  "repository": "git://github.com/Raynos/function-bind.git",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Raynos/function-bind.git"
+  },
+  "funding": {
+    "url": "https://github.com/sponsors/ljharb"
+  },
   "main": "index",
   "homepage": "https://github.com/Raynos/function-bind",
   "contributors": [
@@ -25,24 +31,29 @@
     "url": "https://github.com/Raynos/function-bind/issues",
     "email": "raynos2@gmail.com"
   },
-  "dependencies": {},
   "devDependencies": {
-    "@ljharb/eslint-config": "^12.2.1",
-    "covert": "^1.1.0",
-    "eslint": "^4.5.0",
-    "jscs": "^3.0.7",
-    "tape": "^4.8.0"
+    "@ljharb/eslint-config": "^21.1.0",
+    "aud": "^2.0.3",
+    "auto-changelog": "^2.4.0",
+    "eslint": "=8.8.0",
+    "in-publish": "^2.0.1",
+    "npmignore": "^0.3.0",
+    "nyc": "^10.3.2",
+    "safe-publish-latest": "^2.0.0",
+    "tape": "^5.7.1"
   },
   "license": "MIT",
   "scripts": {
+    "prepublishOnly": "safe-publish-latest",
+    "prepublish": "not-in-publish || npm run prepublishOnly",
+    "prepack": "npmignore --auto --commentLines=autogenerated",
     "pretest": "npm run lint",
     "test": "npm run tests-only",
-    "posttest": "npm run coverage -- --quiet",
-    "tests-only": "node test",
-    "coverage": "covert test/*.js",
-    "lint": "npm run jscs && npm run eslint",
-    "jscs": "jscs *.js */*.js",
-    "eslint": "eslint *.js */*.js"
+    "posttest": "aud --production",
+    "tests-only": "nyc tape 'test/**/*.js'",
+    "lint": "eslint --ext=js,mjs .",
+    "version": "auto-changelog && git add CHANGELOG.md",
+    "postversion": "auto-changelog && git add CHANGELOG.md && git commit --no-edit --amend && git tag -f \"v$(node -e \"console.log(require('./package.json').version)\")\""
   },
   "testling": {
     "files": "test/index.js",
@@ -58,6 +69,19 @@
       "ipad/6.0..latest",
       "iphone/6.0..latest",
       "android-browser/4.2..latest"
+    ]
+  },
+  "auto-changelog": {
+    "output": "CHANGELOG.md",
+    "template": "keepachangelog",
+    "unreleased": false,
+    "commitLimit": false,
+    "backfillLimit": false,
+    "hideCredit": true
+  },
+  "publishConfig": {
+    "ignore": [
+      ".github/workflows"
     ]
   }
 }
