@@ -25,7 +25,7 @@ userSchema.pre('save', async function (next) {
         next()
     }
     const salt = await bcrypt.genSalt(10);
-    this.password = await bcrypt.hash(this.password, salt)
+    this.password =  bcrypt.hash(this.password, salt)
 
 })
 const User = mongoose.model("User", userSchema);
