@@ -33,7 +33,7 @@ app.use(errorHandler)
 
 
 
-const PORT = process.env.PORT || 5000
+const PORT = process.env.PORT
 
 const server = app.listen(5000, console.log(`server started on port ${PORT}`.blue.bold))
 const io = require("socket.io")(server, {
