@@ -123,7 +123,7 @@ const SingleChat = ({ fetchAgain, setFetchAgain }) => {
       );
       setMessages(data);
       setLoading(false);
-      console.log(messages);
+     
 
       socket.emit("join chat", selectedChat._id);
     } catch (error) {
