@@ -17,7 +17,7 @@ const SignIn = () => {
         setLoading(true);
         if (!email || !password) {
             toast({
-                title: "Please Fill all the Feilds",
+                title: "Please Fill all the Fields",
                 status: "warning",
                 duration: 5000,
                 isClosable: true,
