@@ -29,7 +29,7 @@ const SignUp = () => {
             });
             return;
         }
-        console.log(pics);
+     //   console.log(pics);
         if (pics.type === "image/jpeg" || pics.type === "image/png") {
             const data = new FormData();
             data.append("file", pics);
@@ -42,11 +42,11 @@ const SignUp = () => {
                 .then((res) => res.json())
                 .then((data) => {
                     setPic(data.url.toString());
-                    console.log(data.url.toString());
+                    //console.log(data.url.toString());
                     setLoading(false);
                 })
                 .catch((err) => {
-                    console.log(err);
+                   // console.log(err);
                     setLoading(false);
                 });
         } else {
@@ -87,7 +87,7 @@ const SignUp = () => {
             })
             return;
         }
-        console.log(name, email, password, pic);
+        //console.log(name, email, password, pic);
         try {
             const config = {
                 headers: {
